Refactor common setup and cleanup code in satellite process tests.
